TST: collect CategoricalIndex tests by method
